training-model-pipeline/requirements.txt
apache-airflow>=2.7.1
python-dotenv>=1.0.0
mlflow>=2.8.1
scikit-learn>=1.3.2
pandas>=2.1.3
numpy>=1.24.3
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0
python-dotenv>=1.0.0
dvc>=3.30.1 